requirements.txt
PyQt5==5.15.9
python-dotenv==1.0.0
secure-smtplib==0.1.1 